rename folder to follow MVC practices
