+++ application/index.js
@@ -5,8 +5,8 @@ const defaultPort = 3000;
 const PORT = process.env.PORT || defaultPort;
 
 const app = express();
-app.use('/', require('./routes/index'));
+app.use('/', require('./controllers/index'));
 
 app.listen(PORT, () => {
-    console.log(`Server started listening on PORT ${PORT}`);
+    console.log(`Server started listening on PORT ${PORT}. Visit http://localhost:3000 in your browser`);
 });
